+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 4.1.0
+
 * Adding custom_parameters arg to ElastiCache to set custom parameters
 
 # 4.0.1